Don't limit the results when there's a dedup
There was an issue when we limited the results. If you scrolled down to
get more results, then it would run the query again with later data to
add, but those results didn't take into account the existing data,
meaning you could get duplicated field values for your dedup field after
scrolling. To address this, we just run dedup on all the records instead
of a subset, so that all the results are sent to the UI and so scrolling
down doesn't rerun the query.