Spl dedup (#375)
* Rename Statistic Expr Options attribute

* Allow parsing spl dedup cmd

* Add a part of processing logic for dedup cmd

* fixups

* fixup

* Convert DedupComb to a map

* Cleanup

* Don't limit the results when there's a dedup

There was an issue when we limited the results. If you scrolled down to
get more results, then it would run the query again with later data to
add, but those results didn't take into account the existing data,
meaning you could get duplicated field values for your dedup field after
scrolling. To address this, we just run dedup on all the records instead
of a subset, so that all the results are sent to the UI and so scrolling
down doesn't rerun the query.

* Fix issue with scrolling in UI for dedup queries

* Cleanup

* Add unit tests

* Cleanup

* Fix broken unit tests

Some of the unit tests were broken because GetJsonFromAllRrc wasn't
necessarily preserving the order of allrrc in the list of records that
it returned.

For example, Test_GetJsonFromAllRrc_withAggs_ExcludeCols() would
sometimes get the records in the wrong order, so the test would fail.

---------

Co-authored-by: tcefrep <1626680964@qq.com>