Fix issue counting segments
Recently we switched from doing cleaning on each segment to doing it on
each block within the segment, but some parts still assumed that every
time they were called it was on a new segment. This messed up some logic
for things that need to have all the results before they can return
data, like dedup.