Fix issue with regex when string is empty
When the string to compile is empty, we used to compile it to regex
anyways and get an empty (but not nil) *Regexp (and no error). So there
wasn't a way to indicate that regex should not be used in that case. Now
when this happens we get a nil *Regexp, indicating that we should not
use regex.