Fix json encoding (#775)
The `key` here is a string, but it's not really a string. Really it's
meant to be a []byte, but we use it as a string so that it can be a key
in a map. This causes problems during the encoding and decoding process
for values in the []byte that fall outside the normal ASCII range. This
changes fixes the issue by doing a base64 encoding of the `key` so that
all characters are ASCII characters, and then we can use normal JSON
encoding for that string. Without this, the reconstructed value differed
from the origin value that was encoded.