Parallelize search to improve bottleneck query speed (#2748)
* Add unit test for if search can be parallelized

* Fix failing unit test

* Use anyOrder sorting in searcher when possible

* Extract helper function

* Revert "Extract helper function"

This reverts commit 3ebd4b61d9aa72868300c79e2533b0d7e24502ea.

* Parallelize fetching unsorted RRCs

* Fix lint

* Refactor

* Cleanup cached stream initialization

* Create the parallel chain

* fixup

* Fetch DP streams in parallel

* Make TwoWayMap thread safe

* fixups

* Fix bug not unlocking mutex

* Init map with correct size

* Cleanup

* Cleanup

* Update unit test

* Fix merging into correct DP

* Avoid parallelizing data generators

We can parallelize them later, but right now we avoid it because
multithreaded calls to them aren't handled properly so it generates
duplicate data.

* Delay lock release to avoid race condition

Each batch refers to the same NodeResult, which I believe is what was
causing a race condition sometimes. e.g., for:
  app_name=Companyhave
  | fields city, bool_col, http_status, app_name, country
  | sort http_status
  | stats count
on the functional dataset.

* Fixup for connecting parallelized DPs to merge DP

* Fixup for queries with stats as first agg

These queries go through a slightly different flow. We should probably
unify this at some point so it doesn't matter if stats is the first or
n-th agg. For now, this fix avoids a panic.

* Fix issue when first agg is stats

* Refactor

* Add single-threaded stream to fix race condition

* Cleanup