Reuse first chain to avoid mangling
asDataProcessor() has side-effects for top/rare; when the conversion
from the aggs linked list to the DP chain was done multiple times for
the query search | stats | top, the second conversion was mangled
because the first conversion changed the linked list.

An alternative approach would be to do the conversion once and deep copy
it as needed, but there is a fair bit of code needed to add deep copying
methods in all the necessary places.