Reapply #2748 with fixes (#2764)
* Reapply "Parallelize search to improve bottleneck query speed (#2748)" (#2761)

This reverts commit 89216b93df0d554f54b153ef1d74ff0868916814.

* Avoid panic

Some flows can change the dataProcessors list; this change keeps track
of the original one to avoid a potential panic (access past slice end).

* Fix mangled DP chain for queries with top/rare

* Reuse first chain to avoid mangling

asDataProcessor() has side-effects for top/rare; when the conversion
from the aggs linked list to the DP chain was done multiple times for
the query search | stats | top, the second conversion was mangled
because the first conversion changed the linked list.

An alternative approach would be to do the conversion once and deep copy
it as needed, but there is a fair bit of code needed to add deep copying
methods in all the necessary places.